S0C-3A-3A: Refactor code structure for improved readability and maintainability
